riemann: extract RiemannDestWorker into a separate compilation unit
